working on cleaning up inheritance
